[PM-24468] Move HIBP logic to its own module
