+++ crates/bitwarden-vault/src/cipher_risk/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod types;
 pub(crate) mod cipher_risk_client;
+mod hibp;
 
 pub use types::{
     CipherLoginDetails, CipherRisk, CipherRiskOptions, ExposedPasswordResult, PasswordReuseMap,
